fix: move datagpt
