+++ apps/app/pages/datagpt.tsx
@@ -1,5 +1,4 @@
-import CatalogueDataGPT from "@data-catalogue/datagpt";
-import { get } from "datagovmy-ui/api";
+import DataGPT from "@misc/datagpt";
 import { Metadata, Progress } from "datagovmy-ui/components";
 import { withi18n } from "datagovmy-ui/decorators";
 import { useTranslation } from "datagovmy-ui/hooks";
@@ -7,22 +6,22 @@ import { Page } from "datagovmy-ui/types";
 import { GetStaticProps, InferGetStaticPropsType } from "next";
 
 const AIChat: Page = ({}: InferGetStaticPropsType<typeof getStaticProps>) => {
-  const { t } = useTranslation();
+  const { t } = useTranslation(["datagpt"]);
 
   return (
     <>
       <Metadata title={"DataGPT"} description={t("description")} keywords={""} />
       <Progress />
-      <CatalogueDataGPT />
+      <DataGPT />
     </>
   );
 };
 
-export const getStaticProps: GetStaticProps = withi18n("catalogue-datagpt", async () => {
+export const getStaticProps: GetStaticProps = withi18n("datagpt", async () => {
   return {
     props: {
       meta: {
-        id: "catalogue-datagpt",
+        id: "datagpt",
         type: "misc",
         category: null,
         agency: null,
